cmp_fields as a float to work with memory

+++ src/util/sort_fields.h
@@ -7,4 +7,4 @@
 proc_t *sort_by_field(proc_t *procs, int field, int nproc);
 
 void cur_fields(proc_t *proc_arr[], proc_t *cur, int proc_index,
-                int field, uint64_t cmp_fields[]);
+                int field, float cmp_fields[]);
